backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.8
httpx==0.27.2
tenacity==9.0.0
pydantic==2.9.2
pytz==2024.2
structlog==24.1.0
beautifulsoup4==4.12.3
pytest==8.3.3
google-generativeai==0.8.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1

